fix: missing this context (#12)
* v1.0.3-beta

* v1.0.3-beta1

* v1.0.3-beta2

* v1.0.3-beta3

* v1.0.3-beta4

* fix: update Babel to transpile back to Node 6

* fix: update to string-based docs for schema

* fix: add model to an object

Because `context` gets copied and combined with `extraContext`, we need
to export an object rather than an instance of a class. Otherwise we get
issues with `this` misbehaving.

* chore: upgrade dependencies

* test: fix Babel coverage with temporary plugin

Until istanbuljs/istanbuljs#125 is merged, we need a temporary plugin to
get coverage to report correctly.

BREAKING CHANGE: Moves model class from `context` prop directly to a `context.model` prop, which prevents issues when merging context objects.